// src/.generated/templates/Feature.www.model.ts
/**
 * <auto-generated>
 *     This code was generated by a tool.
 *
 *     Changes to this file may cause incorrect behavior and will be lost if the code is regenerated.
 * </auto-generated>
 */ 

// @ts-ignore Prevent noUnusedLocals from triggering an error, we import all core field types to minimize changing to code generation.
import { Sitecore } from "./_.Sitecore.Override"
// @ts-ignore Prevent noUnusedLocals from triggering an error, we import all core field types to minimize changing to code generation.
import { ComponentRendering, RouteData, Field, ImageField, FileField, LinkField } from '@sitecore-jss/sitecore-jss-nextjs';
// @ts-ignore Prevent noUnusedLocals from triggering an error, we import all core field types to minimize changing to code generation.
import { DefaultComponentProps } from '@/lib/component-props';


export namespace Feature.www.Sitecore.templates.Feature.www.Tiles.Fields {
    export type RenderingParameter = { 
        fields?: {  
        }
 }; 
}

export namespace Feature.www.Sitecore.templates.Feature.www.Tiles.Fields {
    export type Tile = { 
        fields?: { 
            /**
            * Represents the Attractions field (c3123540-1c81-4064-a409-dc465d831aec).
            */
            Attractions: Sitecore.Override.ItemEx[];

            /**
            * Represents the Destinations field (d436064f-dcff-436c-9fe7-55355b8938aa).
            */
            Destinations: Sitecore.Override.ItemEx[];

            /**
            * Represents the Type field (0537801f-f977-4b6b-a017-753f522efd4e).
            */
            Type: Field<string>;
 
        }
 }; 
}

export namespace Feature.www.Sitecore.templates.Feature.www.Tiles.Fields {
    export type TileFolder = { 
        fields?: {  
        }
 }; 
}

export namespace Feature.www.Sitecore.templates.Feature.www.Navigation.Fields {
    export type Wwwnavigation = { 
        fields?: { 
            /**
            * Represents the HeaderLinks field (808ed3c8-5469-4073-900d-ce104fbf4672).
            */
            HeaderLinks: Sitecore.Override.ItemEx[];

            /**
            * Represents the HeaderLogo field (b95a6c5b-cc24-4143-b0d3-23e8cab1e88a).
            */
            HeaderLogo: ImageField;

            /**
            * Represents the HeaderText field (037468f4-af0c-4b02-9350-4a1aebfd9c80).
            */
            HeaderText: Field<string>;
 
        }
 }; 
}

export namespace Feature.www.Sitecore.templates.Feature.www.Navigation.RenderingParameters.Fields {
    export type Wwwnavigation = { 
        fields?: {  
        }
 }; 
}

export namespace Feature.www.Sitecore.templates.Feature.www.Navigation.Fields {
    export type WwwnavigationFolder = { 
        fields?: {  
        }
 }; 
}

export namespace Feature.www.Sitecore.templates.Feature.www.Logo.Fields {
    export type logo = { 
        fields?: { 
            /**
            * Represents the Image field (8dace9e1-958d-44c2-b0d4-076438d5425b).
            */
            Image: ImageField;

            /**
            * Represents the Caption field (bd26c8b4-93dd-4f9d-9073-c7d364a462c9).
            */
            ImageCaption: Field<string>;

            /**
            * Represents the URL field (c03f67b5-75ed-4d58-9281-bfa6ee6345f9).
            */
            TargetUrl: LinkField;
 
        }
 }; 
}

export namespace Feature.www.Sitecore.templates.Feature.www.Logo.RenderingParameters.Fields {
    export type logo = { 
        fields?: {  
        }
 }; 
}

export namespace Feature.www.Sitecore.templates.Feature.www.Logo.Fields {
    export type logoFolder = { 
        fields?: {  
        }
 }; 
}